// src/types/Connector.ts
export type Trigger = {
  key: string;
  name: string;
  display: Display;
  operation?: TriggerOperation; // TODO: make required
  inputFields?: Field[];
};

export type TriggerOperation =
  | ChainEventOperation
  | HookOperation
  | PollingOperation;

export type ChainEventOperation = {
  type: string; //"blockchain:event"; // Can't use literals in JSON file: https://github.com/microsoft/TypeScript/issues/26552
  chains: Chain[];
  signature: string;
  filters: ChainEventOperationFilter;
  inputFields?: Field[];
  inputFieldProviderUrl?: string;
  outputFields?: Field[];
  sample?: { [key: string]: string | number | boolean | (string | number)[] };
};

export type ChainEventOperationFilter = {
  fromBlock?: string | number;
  toBlock?: string | number;
  address?: string | string[];
  topics?: string[];
};

export type HookOperation = {
  type: string; //"hook";
  inputFields?: Field[];
  inputFieldProviderUrl?: string;
  outputFields?: Field[];
  sample?: { [key: string]: string | number | boolean | (string | number)[] };
};

export type PollingOperation = {
  type: string; //"polling";
  operation?: Request; // TODO: make required
  inputFields?: Field[];
  inputFieldProviderUrl?: string;
  outputFields?: Field[];
  sample?: { [key: string]: string | number | boolean | (string | number)[] };
};

// End Triggers

// Actions

export type Action = {
  key: string;
  name: string;
  display: Display;
  operation?: ActionOperation; // TODO: make required
  inputFields?: Field[];
};

export type ActionOperation = ChainCallOperation | APICallOperation;

export type APICallOperation = {
  type: string; //"api";
  operation: Request;
  inputFields?: Field[];
  inputFieldProviderUrl?: string;
  outputFields?: Field[];
  sample?: { [key: string]: string | number | boolean | (string | number)[] };
};

export type ChainCallOperation = {
  type: string; //"blockchain:call";
  accounts?: ChainAccount[];
  signature?: string; // TODO: make required
  arguments?: ChainCallOperationArgs[]; // TODO: make required
  inputFields?: Field[];
  inputFieldProviderUrl?: string;
  outputFields?: Field[];
  sample?: { [key: string]: string | number | boolean | (string | number)[] };
};
// ...
export type Display = {
  label: string;
  description: string;
  instructions?: string;
  icon?: string;
  featured?: boolean;
};

export type Field = {
  key: string;
  label?: string;
  helpText?: string;
  type?: string;
  /*| "string"
      | "text"
      | "integer"
      | "number"
      | "boolean"
      | "datetime"
      | "file"
      | "password"
      | "copy"
      | "code";*/
  required?: boolean;
  placeholder?: string;
  default?: string;
  choices?: FieldChoice[];
  list?: boolean;
  children?: Field[];
  dict?: boolean;
  computed?: boolean;
  updateFieldDefinition?: boolean;
  inputFormat?: string;
  validation?: any;
  readonly?: boolean;
};

export type FieldChoice =
  | string
  | {
      value: string;
      label: string;
      sample: string;
      icon?: string;
    };